Fixed bugs related to Proto formation
From Stephen:
Is this a bug?
                    ```agobj.sugar_level = agobj.sugar_level - threshold
                    transaction1 = Transaction(agobj.sugar_level - threshold,
                                               timeperiod, "deposit", 
                                               agobj.agent_id)```

Let's say the agent's sugar level is 84 going in, and the threshold is 50. It seems to me that first, its sugar level will be decremented to 34, and then, a transaction for a (negative) deposit of -16 is created, leaving the new proto with a negative balance.

-----------------------------------------
One other possible bug (but please tell me if it's not a bug and instead I'm misreading things) is the part where a new Proto object is created. It assigns the new Proto an id of start_proto_id, which is fine so far, but it doesn't increment the value of start_proto_id inside the loop, so if there is ever more than one new Proto object created in a single call to form_possible_protos!(), they'll all get the same proto_id value.

Again, tell me if I'm wrong. I don't want to break anything by "fixing" it!